backend/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
PyMuPDF
Pillow
numpy
TTS
torch